refactor(matrix-filter): move derived types from constants to types file
